+++ HISTORY.rst
@@ -11,6 +11,7 @@ Historia zmian
 * raport ewaluacja - upoważnienia (#1083),
 * sprawdzanie i ostrzeganie użytkownika przy zapisie rekordów w sytuacji, gdy dane DOI lub WWW
   już istnieją w bazie danych (#1059),
+* raport rozbieżności autor-źródło (#1023),
 
 202110.1081-rc0
 ---------------